requirements.txt
streamlit
pandas
python-dotenv
reportlab
scikit-learn
matplotlib
requests
folium
gspread
gspread-dataframe
google-auth
openai
prophet
statsmodels
python-pptx
torch
tqdm
scipy
oauth2client
time
